+++ src/main/java/com/atlauncher/gui/tabs/tools/ToolsViewModel.java
@@ -21,11 +21,11 @@ import java.io.File;
 import java.net.InetAddress;
 import java.nio.file.Files;
 import java.util.Arrays;
+import java.util.Locale;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
 import com.atlauncher.App;
-import com.atlauncher.Data;
 import com.atlauncher.FileSystem;
 import com.atlauncher.constants.Constants;
 import com.atlauncher.evnt.listener.SettingsListener;
@@ -130,7 +130,8 @@ public class ToolsViewModel implements IToolsViewModel, SettingsListener {
 
     private final String[] HOSTS = { "authserver.mojang.com", "session.minecraft.net", "libraries.minecraft.net",
             "launchermeta.mojang.com", "launcher.mojang.com", Constants.API_HOST, Constants.PASTE_HOST,
-            Constants.DOWNLOAD_HOST, Constants.FABRIC_HOST, Constants.LEGACY_FABRIC_HOST, Constants.NEOFORGE_HOST, Constants.FORGE_HOST,
+            Constants.DOWNLOAD_HOST, Constants.FABRIC_HOST, Constants.LEGACY_FABRIC_HOST, Constants.NEOFORGE_HOST,
+            Constants.FORGE_HOST,
             Constants.QUILT_HOST, Constants.CURSEFORGE_CORE_API_HOST, Constants.MODRINTH_HOST,
             Constants.MODPACKS_CH_HOST };
 
@@ -254,7 +255,7 @@ public class ToolsViewModel implements IToolsViewModel, SettingsListener {
 
         // Response Code Test
         try {
-            results.append(String.format("Response code to %s was %d\n\n----------------\n\n",
+            results.append(String.format(Locale.ENGLISH, "Response code to %s was %d\n\n----------------\n\n",
                     Constants.DOWNLOAD_SERVER,
                     Download.build()
                             .setUrl(String.format("%s/launcher/json/files.json", Constants.DOWNLOAD_SERVER))
@@ -293,10 +294,12 @@ public class ToolsViewModel implements IToolsViewModel, SettingsListener {
         float kbps = bps / 1024;
         float mbps = kbps / 1024;
         String speed = (mbps < 1
-                ? (kbps < 1 ? String.format("%.2f B/s", bps) : String.format("%.2f " + "KB/s", kbps))
-                : String.format("%.2f MB/s", mbps));
+                ? (kbps < 1 ? String.format(Locale.ENGLISH, "%.2f B/s", bps)
+                        : String.format(Locale.ENGLISH, "%.2f KB/s", kbps))
+                : String.format(Locale.ENGLISH, "%.2f MB/s", mbps));
         results.append(
-                String.format("Download speed to %s was %s, " + "" + "taking %.2f seconds to download 100MB",
+                String.format(Locale.ENGLISH,
+                        "Download speed to %s was %s, taking %.2f seconds to download 100MB",
                         Constants.DOWNLOAD_SERVER, speed, (timeTaken / 1000.0)));
         onTaskComplete.accept(null);