+++ src/main/java/com/atlauncher/utils/CurseForgeApi.java
@@ -25,6 +25,7 @@ import java.util.Arrays;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
@@ -69,7 +70,7 @@ public class CurseForgeApi {
     public static List<CurseForgeProject> searchCurseForge(String gameVersion, int sectionId, String query, int page,
             List<Integer> modLoaderTypes, String sort, boolean sortDescending, Integer categoryId) {
         try {
-            String url = String.format(
+            String url = String.format(Locale.ENGLISH,
                     "%s/mods/search?gameId=432&classId=%s&searchFilter=%s&sortField=%s&sortOrder=%s&pageSize=%d&index=%d",
                     Constants.CURSEFORGE_CORE_API_URL, sectionId,
                     URLEncoder.encode(query, StandardCharsets.UTF_8.name()),
@@ -198,7 +199,8 @@ public class CurseForgeApi {
     }
 
     public static List<CurseForgeFile> getFilesForProject(int projectId) {
-        String url = String.format("%s/mods/%d/files?pageSize=1000", Constants.CURSEFORGE_CORE_API_URL, projectId);
+        String url = String.format(Locale.ENGLISH, "%s/mods/%d/files?pageSize=1000", Constants.CURSEFORGE_CORE_API_URL,
+                projectId);
 
         Download download = Download.build().setUrl(url).header("x-api-key", Constants.CURSEFORGE_CORE_API_KEY)
                 .cached(new CacheControl.Builder().maxStale(10, TimeUnit.MINUTES).build());
@@ -216,7 +218,8 @@ public class CurseForgeApi {
     }
 
     public static CurseForgeFile getFileForProject(int projectId, int fileId) {
-        String url = String.format("%s/mods/%d/files/%d", Constants.CURSEFORGE_CORE_API_URL, projectId, fileId);
+        String url = String.format(Locale.ENGLISH, "%s/mods/%d/files/%d", Constants.CURSEFORGE_CORE_API_URL, projectId,
+                fileId);
 
         Download download = Download.build().setUrl(url).header("x-api-key", Constants.CURSEFORGE_CORE_API_KEY)
                 .cached(new CacheControl.Builder().maxStale(1, TimeUnit.HOURS).build());
@@ -238,7 +241,7 @@ public class CurseForgeApi {
     }
 
     public static CurseForgeProject getProjectById(String projectId) {
-        String url = String.format("%s/mods/%s", Constants.CURSEFORGE_CORE_API_URL, projectId);
+        String url = String.format(Locale.ENGLISH, "%s/mods/%s", Constants.CURSEFORGE_CORE_API_URL, projectId);
 
         Download download = Download.build().setUrl(url).header("x-api-key", Constants.CURSEFORGE_CORE_API_KEY)
                 .cached(new CacheControl.Builder().maxStale(10, TimeUnit.MINUTES).build());
@@ -264,7 +267,8 @@ public class CurseForgeApi {
     }
 
     private static CurseForgeProject getProjectBySlug(String slug, int classId) {
-        String url = String.format("%s/mods/search?gameId=432&slug=%s&classId=%s", Constants.CURSEFORGE_CORE_API_URL,
+        String url = String.format(Locale.ENGLISH, "%s/mods/search?gameId=432&slug=%s&classId=%s",
+                Constants.CURSEFORGE_CORE_API_URL,
                 slug, classId);
 
         Download download = Download.build().setUrl(url).header("x-api-key", Constants.CURSEFORGE_CORE_API_KEY)